Truly fix NSUInteger formats in unit tests on both platforms.

+++ Tests/NSIndexSetBlocksKitTest.m
@@ -18,8 +18,8 @@
 - (void)testEach {
 	NSMutableString *order = [NSMutableString string];
 	BKIndexBlock indexBlock = ^(NSUInteger index) {
-		[order appendFormat:@"%ld", index];
-		_target[index] = [NSString stringWithFormat:@"%ld", index];
+		[order appendFormat:@"%lu", (unsigned long)index];
+		_target[index] = [NSString stringWithFormat:@"%lu", (unsigned long)index];
 	};
 	[_subject each: indexBlock];
 	STAssertTrue([order isEqualToString: @"123"], @"the index loop order is %@", order);
@@ -30,10 +30,10 @@
 - (void)testMatch {
 	NSMutableString *order = [NSMutableString string];
 	BKIndexValidationBlock indexValidationBlock = ^(NSUInteger index) {
-		[order appendFormat:@"%ld", index];
+		[order appendFormat:@"%lu", (unsigned long)index];
 		BOOL match = NO;
 		if (index%2 == 0 ) {
-			_target[index] = [NSString stringWithFormat:@"%ld",index];
+			_target[index] = [NSString stringWithFormat:@"%lu", (unsigned long)index];
 			match = YES;
 		}
 		return match;
@@ -46,7 +46,7 @@
 - (void)testNotMatch {
 	NSMutableString *order = [NSMutableString string];
 	BKIndexValidationBlock indexValidationBlock = ^(NSUInteger index) {
-		[order appendFormat: @"%ld", index];
+		[order appendFormat: @"%lu", (unsigned long)index];
 		BOOL match = index > 4 ? YES : NO;
 		return match;
 	};
@@ -58,7 +58,7 @@
 - (void)testSelect {
 	NSMutableString *order = [NSMutableString string];
 	BKIndexValidationBlock indexValidationBlock = ^(NSUInteger index) {
-		[order appendFormat:@"%ld", index];
+		[order appendFormat:@"%lu", (unsigned long)index];
 		BOOL match = index < 3 ? YES : NO;
 		return match;
 	};
@@ -71,7 +71,7 @@
 - (void)testSelectedNone {
 	NSMutableString *order = [NSMutableString string];
 	BKIndexValidationBlock indexValidationBlock = ^(NSUInteger index) {
-		[order appendFormat:@"%ld", index];
+		[order appendFormat:@"%lu", (unsigned long)index];
 		BOOL match = index == 0 ? YES : NO;
 		return match;
 	};
@@ -83,7 +83,7 @@
 - (void)testReject {
 	NSMutableString *order = [NSMutableString string];
 	BKIndexValidationBlock indexValidationBlock = ^(NSUInteger index) {
-		[order appendFormat:@"%ld",index];
+		[order appendFormat:@"%lu", (unsigned long)index];
 		BOOL match = [[_target objectAtIndex:index] isEqual: @"0"] ? YES : NO;
 		return match;
 	};
@@ -95,7 +95,7 @@
 - (void)testRejectedNone {
 	NSMutableString *order = [NSMutableString string];
 	BKIndexValidationBlock indexValidationBlock = ^(NSUInteger index) {
-		[order appendFormat:@"%ld",index];
+		[order appendFormat:@"%lu", (unsigned long)index];
 		BOOL match = [[_target objectAtIndex:index] isEqual: @"0"] ? NO : YES;
 		return match;
 	};
@@ -107,10 +107,10 @@
 - (void)testAny {
 	__block NSMutableString *order = [NSMutableString string];
 	BKIndexValidationBlock indexValidationBlock = ^(NSUInteger index) {
-		[order appendFormat:@"%ld",index];
+		[order appendFormat:@"%lu", (unsigned long)index];
 		BOOL match = NO;
 		if (index%2 == 0 ) {
-			_target[index] = [NSString stringWithFormat:@"%ld", index];
+			_target[index] = [NSString stringWithFormat:@"%lu", (unsigned long)index];
 			match = YES;
 		}
 		return match;
